aarch64: gicv3: icc regs are serialized in 64bit chunks
The `KVM_DEV_ARM_VGIC_GRP_CPU_SYSREGS` group of GICv3 registers are
serialized in 64bit chunks, not 32bit chunks.

Added a regression unit-test that fails without this fix and passes
with it.

+++ src/vmm/src/vstate/vm.rs
@@ -347,7 +347,7 @@ pub struct VmState {
 pub struct VmState {
     dist: Vec<u32>,
     rdist: Vec<u32>,
-    icc: Vec<u32>,
+    icc: Vec<u64>,
 }
 
 #[cfg(test)]
